# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_smart-invest-solutions_backend_internal_domain.CreateLifeInsuranceDTO:
    properties:
      company_name:
        type: string
      family_member_id:
        type: string
      is_mapped:
        type: boolean
      policy_details:
        $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.CreatePolicyDetailsDTO'
      premium_details:
        $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.CreatePremiumDetailsDTO'
      user_id:
        example: 64f1a2b3c4d5e6f7a8b9c0d1
        type: string
    required:
    - company_name
    - family_member_id
    - policy_details
    - premium_details
    type: object
  github_com_smart-invest-solutions_backend_internal_domain.CreatePolicyDetailsDTO:
    properties:
      doc:
        type: string
      maturity_date:
        type: string
      nominee_name:
        type: string
      plan_name:
        type: string
      policy_no:
        type: string
      ppt:
        type: integer
      sum_assured:
        type: number
      term:
        type: integer
    required:
    - doc
    - maturity_date
    - nominee_name
    - plan_name
    - policy_no
    - ppt
    - sum_assured
    - term
    type: object
  github_com_smart-invest-solutions_backend_internal_domain.CreatePremiumDetailsDTO:
    properties:
      installment_premium:
        type: number
      next_due_date:
        type: string
      payment_mode:
        enum:
        - Yearly
        - Half-Yearly
        - Quarterly
        - Monthly
        type: string
    required:
    - installment_premium
    - next_due_date
    - payment_mode
    type: object
...
  github_com_smart-invest-solutions_backend_internal_domain.LifeInsurance:
    properties:
      company_name:
        type: string
      created_at:
        type: string
      family_member_id:
        type: string
      id:
        type: string
      is_mapped:
        description: 'Admin flag: formally mapped to their agency portfolio'
        type: boolean
      policy_details:
        $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.PolicyDetails'
      premium_details:
        $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.PremiumDetails'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_smart-invest-solutions_backend_internal_domain.LifeInsuranceListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.LifeInsurance'
        type: array
      total:
        type: integer
    type: object
...
  github_com_smart-invest-solutions_backend_internal_domain.PolicyDetails:
    properties:
      doc:
        description: Date of Commencement
        type: string
      life_insured_name:
        description: Cached from the linked family member
        type: string
      maturity_date:
        type: string
      nominee_name:
        type: string
      plan_name:
        type: string
      policy_no:
        type: string
      ppt:
        description: Premium Paying Term in years
        type: integer
      sum_assured:
        type: number
      term:
        description: Total policy duration in years
        type: integer
    type: object
...
  github_com_smart-invest-solutions_backend_internal_domain.PremiumDetails:
    properties:
      installment_premium:
        type: number
      next_due_date:
        description: Indexed — powers premium-reminder/dashboard queries
        type: string
      payment_mode:
        type: string
    type: object
...
  github_com_smart-invest-solutions_backend_internal_domain.UpdateLifeInsuranceDTO:
    properties:
      company_name:
        type: string
      doc:
        type: string
      family_member_id:
        type: string
      installment_premium:
        type: number
      is_mapped:
        type: boolean
      maturity_date:
        type: string
      next_due_date:
        type: string
      nominee_name:
        type: string
      payment_mode:
        enum:
        - Yearly
        - Half-Yearly
        - Quarterly
        - Monthly
        type: string
      plan_name:
        type: string
      policy_no:
        type: string
      ppt:
        type: integer
      sum_assured:
        type: number
      term:
        type: integer
    type: object
...
  github_com_smart-invest-solutions_backend_pkg_response.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: support@ulmind.com
    name: ULMIND Team
  description: |-
    Backend API for Smart Invest Solutions — a LIC Policy & Investment Management Platform.
    This API provides user registration, authentication with JWT, role-based access control, and more.
  title: Smart Invest Solutions API
  version: "1.0"
paths:
...
  /life-insurances:
    get:
      consumes:
      - application/json
      description: Clients receive their own policies (unpaginated list + total).
        A super_admin receives a paginated master list across all clients (page, limit,
        is_mapped query params); a plain admin receives one scoped to clients under
        their own Agency ID. Each row is enriched with the customer's name, contact
        number, and Agency ID.
      parameters:
      - description: 'Page number — Admin only (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page — Admin only (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: Filter by mapped status — Admin only
        in: query
        name: is_mapped
        type: boolean
      - description: Filter to policies whose insured family member carries this exact
          LIC Customer ID — Admin only
        in: query
        name: lic_customer_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Policies retrieved successfully
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Life Insurance policies
      tags:
      - Life Insurance
    post:
      consumes:
      - application/json
      description: Creates a new life insurance policy mapped to a specific family
        member. Clients create policies under their own account; admin/super_admin
        may target any client by passing user_id in the body ("Select Member" flow).
      parameters:
      - description: Life Insurance Policy Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.CreateLifeInsuranceDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Life Insurance policy added successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.LifeInsurance'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Add Life Insurance policy
      tags:
      - Life Insurance
  /life-insurances/{id}:
    delete:
      consumes:
      - application/json
      description: Removes a life insurance policy. Clients can only delete their
        own; admin/super_admin can delete any.
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Policy deleted successfully
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "400":
          description: Bad request or access denied
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete Life Insurance policy
      tags:
      - Life Insurance
    get:
      consumes:
      - application/json
      description: Retrieves detailed info for a single life insurance policy. Clients
        can only access their own; admin/super_admin can access any.
      parameters:
      - description: Life Insurance Policy ID (MongoDB ObjectID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Policy retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.LifeInsurance'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "404":
          description: Policy not found or access denied
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Life Insurance policy by ID
      tags:
      - Life Insurance
    put:
      consumes:
      - application/json
      description: Updates policy/premium fields. Clients can only update their own
        policies; admin/super_admin can update any. Reassigning family_member_id re-validates
        ownership and refreshes the cached life_insured_name.
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.UpdateLifeInsuranceDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Policy updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.LifeInsurance'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "404":
          description: Policy not found or access denied
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update / Edit Life Insurance policy
      tags:
      - Life Insurance
  /life-insurances/user/{userId}:
    get:
      consumes:
      - application/json
      description: Retrieves every life insurance policy belonging to a specific client.
        A super_admin may target any client; a plain admin only one under their own
        Agency ID.
      parameters:
      - description: Target Client User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Policies retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.LifeInsuranceListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "403":
          description: Forbidden — Admin role required
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user's Life Insurance policies (Admin Only)
      tags:
      - Life Insurance (Admin)
...
  /reports/portfolio:
    get:
      consumes:
      - application/json
      description: Generates a PDF summarizing the client's details, family members,
        and Life/Health/General Insurance + Fixed Deposit holdings. Clients always
        receive their own report; admin/super_admin may generate one for any client
        via the user_id query parameter.
      parameters:
      - description: Target client's User ID — Admin/Super Admin only
        in: query
        name: user_id
        type: string
      produces:
      - application/pdf
      responses:
        "200":
          description: Portfolio PDF
          schema:
            type: file
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Download Client Portfolio Report
      tags:
      - Reports
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your Bearer token in the format: **Bearer &lt;token&gt;**'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"